refactor/#384/make from config modular (#640)
