+++ src/blueprint/blueprint.rs
@@ -7,7 +7,7 @@ use derive_setters::Setters;
 use serde_json::Value;
 
 use super::GlobalTimeout;
-use crate::blueprint::server::Server;
+use crate::blueprint::from_config::Server;
 use crate::config::Upstream;
 use crate::lambda::{Expression, Lambda};
