+++ src/blueprint/mod.rs
@@ -4,9 +4,6 @@ mod from_config;
 mod into_schema;
 mod timeout;
 
-// TODO: make it private
-mod server;
-
 pub use blueprint::*;
-pub use server::*;
+pub use from_config::*;
 pub use timeout::GlobalTimeout;
